[CTST-1817][NFC] Clean building warnings of DPCT
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/277755
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>

+++ clang/lib/Tooling/Tooling.cpp
@@ -706,7 +706,7 @@ int ClangTool::run(ToolAction *Action) {
         return -30 /*MigrationErrorInconsistentFileInDatabase*/;
       }
 
-      for (int index = 0; index < SDKIncludePath.size(); index++) {
+      for (size_t index = 0; index < SDKIncludePath.size(); index++) {
         if (SDKIncludePath[index] == '\\') {
           SDKIncludePath[index] = '/';
         }
